Move FCM code into a service
